fixtures, gulp-mocha, tests for loading/creating

+++ src/db/migrations/20160314122531-create-sprint.js
@@ -21,10 +21,10 @@ module.exports = {
         type: Sequelize.STRING
       },
       resource: {
-        type: Sequelize.DECIMAL
+        type: Sequelize.INTEGER
       },
       factor: {
-        type: Sequelize.DECIMAL
+        type: Sequelize.INTEGER
       },
       staffed: {
         type: Sequelize.BOOLEAN
